% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeN.R
\name{computeN}
\alias{computeN}
\title{Sample Size Calculations for Correlates of Risk}
\usage{
computeN(Nrand, tau, taumax, VEtauToTaumax, VE0toTau, risk0, dropoutRisk,
  propCasesWithS)
}
\arguments{
\item{Nrand}{Number of participants randomized to vaccine arm}

\item{tau}{Biomarker sampling timepoint}

\item{taumax}{End of follow-up time period}

\item{VEtauToTaumax}{VE between 'tau' and 'taumax'}

\item{VE0toTau}{VE between 0 and 'tau'}

\item{risk0}{Placebo-group endpoint risk between 'tau' and 'taumax'}

\item{dropoutRisk}{Dropout risk between 0 and 'taumax'}

\item{propCasesWithS}{Proportion of cases with measured S}
}
\value{
List with the following elements:
\itemize{
  \item N: the number of subjects in the vaccine group at risk at tau
  \item nCases: Number of subjects in the vaccine group at-risk at tau with the clinical event (cases) by taumax.
   \item nControls: Number of subjects in the vaccine group at-risk at tau without the clinical event (controls) by taumax.
   \item nCasesWithS: Number of subjects in the vaccine group at-risk at tau with the clinical event (cases) by taumax and with the biomarker measured
}
}
\description{
Calculates projected sample sizes at the design stage for trials assessing biomarkers as correlates of risk (CoRs).
}
\details{
This function calculates projected sample sizes and includes options to account for dropout and incomplete sample storage.
These projected sample sizes can then be used as input values for the computePower function. Calculations use the following
assumptions:
\enumerate{
  \item Failure time \eqn{T} and censoring time \eqn{C} are independent
  \item \eqn{T|Z=0} follows an exponential distribution with rate parameter \eqn{\theta_t} and \eqn{C|Z=0} follows an
     exponential distribution with rate parameter \eqn{\theta_c}
  \item RRtauToTaumax \eqn{ = P(T <= t|T> \tau, Z=1)/P(T <= t|T> \tau, Z=0)} for all \eqn{t} between tau and taumax (this will only approximately hold).
}
}
\examples{
Nrand = 4100
tau = 3.5
taumax = 24
VEtauToTaumax = 0.75
VE0toTau = 0.75/2
risk0 = 0.034
dropoutRisk = 0.1
propCasesWithS = 1
computeN(Nrand, tau, taumax, VEtauToTaumax, VE0toTau, risk0, dropoutRisk, propCasesWithS)

}
\seealso{
\code{\link{computePower}}
}
