% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerCont.R
\name{plotPowerCont}
\alias{plotPowerCont}
\title{Plot Power Curve against CoR Relative Risk for Continuous Biomarkers}
\usage{
plotPowerCont(outComputePower, outDir = NULL, legendText)
}
\arguments{
\item{outComputePower}{List or list of lists containing output from \code{\link{computePower}}, or character vector specifying the file(s) containing \code{\link{computePower}} output.}

\item{outDir}{Character vector specifying path(s) to output file(s), necessary if \code{outComputePower} is a character vector. Default is \code{NULL}.}

\item{legendText}{Character vector specifying the entirety of the legend text. Order of the parameter values must match that of the \code{\link{computePower}} input parameters in order for legend labels to be accurate.}
}
\value{
Plot displaying power vs. CoR relative risk
}
\description{
Plots the power to detect a normally distributed CoR in vaccine recipients against different
values of the CoR effect size, RRc, which is the relative risk per standard deviation increase
for a noise-free biomarker (\eqn{\rho=1}).
}
\details{
The function's plot can be interpreted in conjunction with the output of the \code{\link{plotVElatCont}} function by
matching the CoR relative risk in the two plots and examining power compared to VE. This sheds light on the importance
of overall VE on power and allows correlates of risk results to be interpreted in terms of
potential correlates of efficacy/protection.
}
\examples{

# Example scenario with continuous biomarker, where values of rho are varied

# Set input parameters for computePower function
nCases <- 10
nControls <- 300
nCasesWithS <- 10
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
PlatVElowest <- 0.2
VElowest <- seq(0, VEoverall, len=5)
M <- 22
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.7, 0.4)
biomType <- "continuous"

# Output from computePower function is stored in an object as a list
pwr <- computePower(nCases=nCases, nCasesWithS=nCasesWithS, nControls=nControls,
                    controlCaseRatio=controlCaseRatio, risk0=risk0, VEoverall=VEoverall,
                    PlatVElowest=PlatVElowest, VElowest=VElowest, M=M, alpha=alpha,
                    sigma2obs=sigma2obs, rho=rho, biomType=biomType)

# Set parameters for plotPowerCont function
# outComputePower is a list containing output from the computePower function
outComputePower <- pwr
legendText <- paste0("rho = ", c(1, 0.7, 0.4))
plotPowerCont(outComputePower=outComputePower, legendText=legendText)

\dontrun{
# Output from computePower function is saved in an RData file
computePower(..., saveDir = "myDir", saveFile = "myFile.RData")
# outComputePower is a character string specifying the file containing the computePower output
# outDir is a character string specifying the outComputePower file directory
outComputePower = "myFile"
outDir = "~/myDir"
legendText <- paste0("rho = ", c(1, 0.7, 0.4))
plotPowerCont(outComputePower, outDir=outDir, legendText = legendText)
}

}
\seealso{
\code{\link{computePower}} \code{\link{plotVElatCont}} \code{\link{plotPowerTri}}
}
