% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerTri.R
\name{plotPowerTri}
\alias{plotPowerTri}
\title{Plot Power Curve against CoR Relative Risk for Trichotomous or Binary Biomarkers}
\usage{
plotPowerTri(outComputePower, outDir = NULL, legendText)
}
\arguments{
\item{outComputePower}{List or list of lists containing output from \code{\link{computePower}}, or character vector specifying the file(s) containing \code{\link{computePower}} output.}

\item{outDir}{Character vector specifying path(s) to output file(s), necessary if \code{outComputePower} is a character vector. Default is \code{NULL}.}

\item{legendText}{Character vector specifying the entirety of the legend text. Order of the parameter values must match that of the \code{\link{computePower}} input parameters in order for legend labels to be accurate.}
}
\value{
Plot displaying power vs. CoR relative risk
}
\description{
Plots the power to detect a trichotomous or binary CoR in vaccine recipients against different
values of the CoR effect size, RRt, which is the relative risk in the high biomarker
response subgroup versus the low biomarker response subgroup.
}
\details{
If multiple levels are specified for the biomarker measurement error input parameters (i.e., for \code{Sens}/\code{Spec} or \code{rho}) in the
\code{\link{computePower}} function, only the first level is used to determine the \eqn{RR_t} values that are plotted on the x-axis.
}
\examples{

# Example scenario with trichotomous biomarker, where values of controlCaseRatio are varied

# Set input parameters for computePower function
nCases <- 10
nControls <- 300
nCasesWithS <- 10
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
VElat0 <- seq(0, VEoverall, len=5)
VElat1 <- rep(VEoverall, 5)
Plat0 <- P0 <- 0.2
Plat2 <- P2 <- 0.6
sens <- spec <- 0.8
FP0 <- FN2 <- 0
M <- 50
alpha <- 0.05
biomType <- "trichotomous"

# Output from computePower function is stored in an object as a list
pwr1 <- computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
                     controlCaseRatio=controlCaseRatio, risk0=risk0, VEoverall=VEoverall,
                     Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2, VElat0=VElat0, VElat1=VElat1,
                     M=M, alpha=alpha, spec=spec, FP0=FP0, sens=sens, FN2=FN2, biomType=biomType)

controlCaseRatio <- 3
pwr2 <- computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
                     controlCaseRatio=controlCaseRatio, risk0=risk0, VEoverall=VEoverall,
                     Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2, VElat0=VElat0, VElat1=VElat1,
                     M=M, alpha=alpha, spec=spec, FP0=FP0, sens=sens, FN2=FN2, biomType=biomType)

controlCaseRatio <- 1
pwr3 <- computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
                     controlCaseRatio=controlCaseRatio, risk0=risk0, VEoverall=VEoverall,
                     Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2, VElat0=VElat0, VElat1=VElat1,
                     M=M, alpha=alpha, spec=spec, FP0=FP0, sens=sens, FN2=FN2, biomType=biomType)

# Set parameters for plotPowerTri function
# outComputePower is a list of lists containing outputs from the computePower function
outComputePower <- list(pwr1, pwr2, pwr3)
legendText <- paste0("controls:cases = ", c("5:1","3:1","1:1"))
plotPowerTri(outComputePower=outComputePower, legendText=legendText)

\dontrun{
# outComputePower is a character vector specifying the files containing computePower output
# outDir is a character vector specifying the outComputePower file directories
outComputePower = c("myFile1", "myFile2", "myFile3")
outDir = rep("~/myDir", 3)
legendText <- paste0("controls:cases = ", c("5:1","3:1","1:1"))
plotPowerTri(outComputePower, outDir=outDir, legendText = legendText)
}

}
\seealso{
\code{\link{computePower}} \code{\link{plotPowerCont}}
}
