% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRRgradVE.R
\name{plotRRgradVE}
\alias{plotRRgradVE}
\title{Plot Ratio of Relative Risks for Higher/Lower Latent Subgroups against CoR Effect Size \eqn{RR_t} for Trichotomous Biomarker}
\usage{
plotRRgradVE(outComputePower, outDir = NULL, legendText)
}
\arguments{
\item{outComputePower}{List or list of lists containing output from \code{\link{computePower}}, or character vector specifying the file(s) containing \code{\link{computePower}} output.}

\item{outDir}{Character vector specifying path(s) to output file(s), necessary if \code{outComputePower} is a character vector. Default is \code{NULL}.}

\item{legendText}{Character vector specifying the entirety of the legend text. Order of the parameter values must match that of the \code{\link{computePower}} input parameters in order for legend labels to be accurate.}
}
\value{
Plot displaying ratio of relative risks vs. CoR relative risk effect size
}
\description{
Plots the ratio of relative risks for the higher and lower latent subgroups (\eqn{RRlat2/RRlat0}) versus
the CoR relative risk effect size (\eqn{RR_t = risk_1(2)/risk_1(0)}).
}
\details{
When \eqn{\rho} is varied, this plot shows how the relationship between the CoR effect size RRt and the ratio of latent relative risks RRlat2/RRlat0 changes
for different values of \eqn{\rho}. RRlat2/RRlat0 is a relative vaccine efficacy parameter because RRlat2 = 1 - VElat2 and RRlat0 = 1 - VElat0.
When \eqn{\rho=1}, RRt = RRlat2/RRlat0 such that a CoR in the vaccine group is equivalent to the relative vaccine efficacy parameter,
whereas for imperfectly measured biomarkers with \eqn{\rho<1}, RRt>RRlat2/RRlat0 and the CoR effect size is closer to the null than the relative vaccine efficacy
parameter is.

Note: RRlat2/RRlat0 = \eqn{(risk_1^{lat}(2)/risk_0^{lat}(2)) / (risk_1^{lat}(0)/risk_0^{lat}(0)) = risk_1^{lat}(2) / risk_1^{lat}(0)},
assuming risk_0 is constant for all s* and x*.
}
\examples{

# Example scenario with trichotomous biomarker, where values of rho are varied

# Set input parameters for computePower function
nCases <- 10
nControls <- 300
nCasesWithS <- 10
controlCaseRatio <- 3
VEoverall <- 0.75
risk0 <- 0.034
VElat0 <- seq(0, VEoverall, len=10)
VElat1 <- rep(VEoverall, 10)
Plat0 <- P0 <- 0.2
Plat2 <- P2 <- 0.6
M <- 20
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.7, 0.4)
biomType <- "trichotomous"

# Output from computePower function is stored in an object as a list
pwr <- computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
                    controlCaseRatio=controlCaseRatio, risk0=risk0, VEoverall=VEoverall,
                    Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2, VElat0=VElat0, VElat1=VElat1,
                    M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType)

# Set parameters for plotPowerCont function
# outComputePower is a list containing output from the computePower function
outComputePower <- pwr
legendText <- paste0("rho = ", c(1, 0.7, 0.4))
plotRRgradVE(outComputePower=outComputePower, legendText=legendText)

\dontrun{
# Output from computePower function is saved in an RData file
computePower(..., saveDir = "myDir", saveFile = "myFile.RData")
# outComputePower is a character string specifying the file containing the computePower output
# outDir is a character string specifying the outComputePower file directory
outComputePower = "myFile"
outDir = "~/myDir"
legendText <- paste0("rho = ", c(1, 0.7, 0.4))
plotRRgradVE(outComputePower, outDir=outDir, legendText = legendText)
}

}
\seealso{
\code{\link{computePower}} \code{\link{plotPowerTri}}
}
