% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePower.R
\name{computePower}
\alias{computePower}
\title{Power Calculations for Assessing Biomarkers as Correlates of Risk (CoRs), Accounting for Measurement Error and Treatment Efficacy}
\usage{
computePower(nCases, nControls, nCasesWithS, controlCaseRatio = NULL,
  VEoverall, risk0, VElat0 = seq(0, VEoverall, len = 20),
  VElat1 = rep(VEoverall, 20), VElowest = NULL, Plat0 = 0.2,
  Plat2 = 0.6, P0 = Plat0, P2 = Plat2, PlatVElowest = NULL,
  spec = NULL, FP0 = NULL, sens = NULL, FN2 = NULL, M = 100,
  alpha = 0.05, sigma2obs = 1, rho = 1, biomType = c("continuous",
  "trichotomous", "binary"), cohort = FALSE, p = NULL,
  tpsMethod = c("PL", "ML", "WL"), saveDir = NULL, saveFile = NULL)
}
\arguments{
\item{nCases}{Number of vaccine recipients at-risk at tau with the clinical event by taumax (regardless of whether the biomarker is measured).}

\item{nControls}{Number of vaccine recipients at-risk at tau \emph{without} the clinical event by taumax (regardless of whether the biomarker is measured).}

\item{nCasesWithS}{Number of vaccine recipients at-risk at tau with the clinical event by taumax and with the biomarker measured.}

\item{controlCaseRatio}{Number of controls sampled per case in the vaccine arm.}

\item{VEoverall}{Overall vaccine efficacy.}

\item{risk0}{Placebo-group endpoint rate between time tau and time taumax.}

\item{VElat0}{For trichotomous (or binary) biomarker, grid of VE (vaccine/placebo) values for the lower protected latent subgroup. Each value of \code{VElat0} corresponds to one unique effect size (RR_t).}

\item{VElat1}{For trichotomous biomarker, grid of VE (vaccine/placebo) values for the medium protected latent subgroup. For binary biomarker, specify as any vector of same length as \code{VElat0}; value does not affect function output.}

\item{VElowest}{For continuous bioarker, a vector of values corresponding to the lowest possible value of VE. Typical applications will range \code{VElowest} from 0 to 1 - \code{RRoverall}.}

\item{Plat0}{For trichotomous (or binary) biomarker, prevalence of lower protected latent subgroup (\eqn{P(X = 0)}, where X is the true latent biomarker).}

\item{Plat2}{For trichotomous (or binary) biomarker, prevalence of higher protected latent subgroup (\eqn{P(X = 2)}, where X is the true latent biomarker).}

\item{P0}{For trichotomous (or binary) biomarker, probability of low biomarker response (\eqn{P(S = 0)}, where S is the observed biomarker). If unspecified, this parameter is set to \code{Plat0}.}

\item{P2}{For trichotomous (or binary) biomarker, probability of high biomarker response (\eqn{P(S = 2)}, where S is the observed biomarker). If unspecified, this parameter is set to \code{Plat2}.}

\item{PlatVElowest}{For continuous biomarker, proportion of vaccine recipients with the lowest value of VE.}

\item{spec}{For trichotomous (or binary) biomarker, numeric scalar or vector specifying the specificity (\eqn{P(S=0|X=0)}) of the observed biomarker. Default is \code{NULL}, which indicates 'approach 2' is to be used.}

\item{FP0}{For trichotomous (or binary) biomarker, numeric scalar or vector specifying the low false positive rate (\eqn{P(S=2|X=0)}) of the observed biomarker. Default is \code{NULL}, which indicates 'approach 2' is to be used.}

\item{sens}{For trichotomous (or binary) biomarker, numeric scalar or vector specifying the sensitivity (\eqn{P(S=2|X=2)}) of the observed biomarker. Default is \code{NULL}, which indicates 'approach 2' is to be used.}

\item{FN2}{For trichotomous (or binary) biomarker, numeric scalar or vector specifying the high false negative rate (\eqn{P(S=0|X=2)}) of the observed biomarker. Default is \code{NULL}, which indicates 'approach 2' is to be used.}

\item{M}{Number of simulated clinical trials.}

\item{alpha}{Two-sided Wald test type-I error rate.}

\item{sigma2obs}{For continuous biomarker, or for trichotomous (or binary) biomarker simulated using 'approach 2', the variance of the observed biomarker.}

\item{rho}{For continuous biomarker, or for trichotomous (or binary) biomarker simulated using 'approach 2', a numeric scalar or vector specifying the fraction of protection-relevant variability in the observed biomarker. The first element of this vector should be 1, corresponding to the case of no measurement error.}

\item{biomType}{Type of biomarker that is used. Default is "continuous"; other choices are "trichotomous" and "binary".}

\item{cohort}{Sampling design to be used. Default is \code{FALSE}, specifying case-control sampling design. If \code{TRUE}, case-cohort sampling is used.}

\item{p}{For case-cohort sampling design, probability that a subject will be in the cohort.}

\item{tpsMethod}{Character denoting method for fitting the logistic regression model. Choose from "PL" for pseudo-likelihood (default), "ML" for maximum likelihood, and "WL" for weighted likelihood.}

\item{saveDir}{Character denoting the directory that the function output is to be saved in. Default is \code{NULL}.}

\item{saveFile}{Character denoting the name of the file the function output will be saved in. Output will be saved as an .RData file. Default is \code{NULL}.}
}
\value{
If trichotomous or binary biomarker, list with the following elements:
\itemize{
  \item power: fraction of simulated trials in which the null hypothesis H_0 (expression (14) of the manuscript) is rejected.
  \item RRt: CoR relative risk effect size (risk1(2)/risk1(0))
  \item risk1_2: vaccine-group endpoint risk for high biomarker responses (P(Y=1|X=2))
  \item risk1_0: vaccine-group endpoint risk for low biomarker responses (P(Y=1|X=0))
  \item VElat2: grid of VE (vaccine/placebo) among higher protected latent group
  \item VElat0: grid of VE (vaccine/placebo) among lower protected latent group
  \item Plat2: prevalence of higher protected
  \item Plat0: prevalence of lower protected
  \item P2: probability of high biomarker response
  \item P0: probability of low biomarker response
  \item alphaLat: logit(Y=1|s=0)
  \item betaLat: logit(Y=1|S=2)-logit(Y=1|s=0)
  \item sens: sensitivity (P(S=2|X=2))
  \item spec: specificity (P(S=0|X=0))
  \item FP0: low false positive rate (P(S=2|X=0))
  \item FN2: high false negative rate (P(S=0|X=2))
  \item Ncomplete: total number of subjects at risk at tau, excluding dropouts
  \item nCases: number of observed cases between tau and taumax
  \item nCasesWithS: number of observed cases between tau and taumax with measured S or S*
  \item VEoverall: overall VE
  \item alpha: two-sided Wald test Type 1 error rate
  \item rho: protection-relevant fraction of the variance of S*
  \item controlCaseRatio: ratio of controls to cases in case-control sampling design
  \item risk0: placeb0-group endpoint risk between tau and taumax
}

If continuous biomarker, list with the following elements:
\itemize{
  \item power: fraction of simulated trials in which the null hypothesis H_0 (expression (16) of the manuscript) is rejected.
  \item RRc: CoR relative risk effect size (risk1(s*)/risk1(s*-1))
  \item betaLat: variable in logistic regression model: logit(risk1lat(x*)) = alphaLat + betaLat x*
  \item alphaLat: variable in logistic regression model: logit(risk1lat(x*)) = alphaLat + betaLat x*
  \item PlatVElowest: prevalence of VElowest
  \item VElowest: lowest VE level for true biomarker X* <= nu
  \item sigma2obs: variance of observed biomarker S*
  \item Ncomplete: total number of subjects at risk at tau, excluding dropouts
  \item nCases: number of observed cases between tau and taumax
  \item nCasesWithS: number of observed cases between tau and taumax with measured S*
  \item VEoverall: overall VE
  \item alpha: two-sided Wald test Type 1 error rate
  \item rho: protection-relevant fraction of the variance of S*
  \item controlCaseRatio: ratio of controls to cases in case-control sampling design
  \item risk0: placeb0-group endpoint risk between tau and taumax
}
}
\description{
Performs power calculations for assessing trichotomous, binary, and continuous biomarkers as correlates of risk,
accounting for measurement error and treatment efficacy. Methods are described in [Gilbert, Janes, and Huang (2015).
``Power/Sample Size Calculations for Assessing Correlates of Risk in Clinical Efficacy Trials.'']
}
\details{
If \code{nCases}, \code{nControls}, and \code{nCasesWithS} are vectors, then \code{rho} must be scalar.

To save output in an .RData file, specify the file directory with \code{saveDir} and the name of the file with \code{saveFile}.

This function asuumes the scenario \code{VElat1 = VEoverall}, which yields the formula
\code{VEoverall = (Plat0*VElat0 + Plat2*VElat2)/(Plat0 + Plat2)}.
For fixed values of \code{Plat0} and \code{Plat2}, this links \code{VElat0} to \code{VElat2}
by the formula \code{VElat2 = (VEoverall*(Plat0+Plat2) - Plat0*VElat0)/Plat2},
which is what the function uses to obtain \code{VElat2}.

Following Step 7 in section 3.1 in the manuscript, the measurement error in a trichotomous (or binary) biomarker is accounted for in one of two ways:
  Approach 1 specifies \code{spec}, \code{sens}, \code{FP0}, and \code{FN2} which determine \code{FP1} and \code{FN1} from equations (7) and (8).
  Vector inputs are acceptable and allows for the evaluation of biomarkers with different levels of measurement error.

  Approach 2 specifies \code{sigma2obs} and \code{rho}; this approach assumes the normal measurement error model (9) in the manuscript.
  Specifying \code{spec=NULL}, \code{sens=NULL}, \code{FP0=NULL}, and \code{FN2=NULL} defaults to approach 2, which is what is used in
  illustrations in the manuscript.

Parameters independent of biomarker type and sampling design: nCases, nControls, nCasesWithS, VEoverall, risk0, M, alpha, tpsMethod, saveDir, saveFile

Parameters for trichotomous (or binary) biomarker: \code{VElat0}, \code{VElat1}, \code{Plat0}, \code{Plat2}, \code{P0},
\code{P2}, \code{biomType} = "trichomous" (or "binary")
  \itemize{
    \item Parameters for Approach 1: \code{sens}, \code{spec}, \code{FP0}, \code{FN2}
    \item Parameters for Approach 2: \code{sigma2obs}, \code{rho}
  }

Parameters for continuous biomarker: \code{VElowest}, \code{PlatVElowest}, \code{sigma2obs}, \code{rho}, \code{biomType} = "continuous"

Parameters for case-control sampling design (without-replacement sampling): \code{controlCaseRatio}

Parameters for case-cohort sampling design: \code{cohort=TRUE}, \code{p}
}
\examples{

## Trichotomous biomarker, Approach 1, varying sens and spec ##
## Specify sens, spec, FP0, FN2
nCases <- 32
nControls <- 1000
nCasesWithS <- 32
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
VElat0 <- seq(0, VEoverall, len=20)  # 20 data points for the power curve
VElat1 <- rep(VEoverall, 20)
Plat0 <- 0.2
Plat2 <- 0.6
P0 <- Plat0  # different values of P0 can be set
P2 <- Plat2  # different values of P2 can be set
sens <- spec <- c(1, 0.9, 0.8, 0.7)
FP0 <- FN2 <- rep(0, 4)
M <- 5
alpha <- 0.05
biomType <- "trichotomous"
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             controlCaseRatio=controlCaseRatio, VEoverall=VEoverall, risk0=risk0,
             VElat0=VElat0, VElat1=VElat1, Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2,
             M=M, alpha=alpha, spec=spec, FP0=FP0, sens=sens, FN2=FN2, biomType=biomType)

\dontrun{
## Trichotomous biomarker, Approach 2, varying rho ##
## Specify rho and sigma2obs

nCases <- 32
nControls <- 1000
nCasesWithS <- 32
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
VElat0 <- seq(0, VEoverall, len=20)
VElat1 <- rep(VEoverall, 20)
Plat0 <- 0.2
Plat2 <- 0.6
P0 <- Plat0
P2 <- Plat2
M <- 5
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.9, 0.7, 0.5)
biomType <- "trichotomous"
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             controlCaseRatio=controlCaseRatio, VEoverall=VEoverall, risk0=risk0,
             VElat0=VElat0, VElat1=VElat1, Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2,
             M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType)


## Binary biomarker, Approach 2, varying rho ##
## Plat0 + Plat2 = 1

nCases <- 32
nControls <- 1000
nCasesWithS <- 32
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
VElat0 <- seq(0, VEoverall, len=20)  # 20 data points for the power curve
VElat1 <- rep(0, 20)  # will not be used by function
Plat0 <- 0.2
Plat2 <- 1 - Plat0
P0 <- Plat0
P2 <- Plat2
M <- 5
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.9, 0.7, 0.5)
biomType <- "binary"
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             controlCaseRatio=controlCaseRatio, VEoverall=VEoverall, risk0=risk0,
             VElat0=VElat0, VElat1=VElat1, Plat0=Plat0, Plat2=Plat2, P0=P0, P2=P2,
             M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType)


## Continuous biomarker, varying rho ##

nCases <- 32
nControls <- 1000
nCasesWithS <- 32
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
PlatVElowest <- 0.2
VElowest <- seq(0, VEoverall, len=20)
M <- 5
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.9, 0.7, 0.5)
biomType <- "continuous"
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             controlCaseRatio=controlCaseRatio, VEoverall=VEoverall, risk0=risk0,
             PlatVElowest=PlatVElowest, VElowest=VElowest, M=M, alpha=alpha,
             sigma2obs=sigma2obs, rho=rho, biomType=biomType)


## Continuous biomarker, case-cohort sampling design, varying p ##
nCases <- 32
nControls <- 1000
nCasesWithS <- 32
VEoverall <- 0.75
risk0 <- 0.034
PlatVElowest <- 0.2
VElowest <- seq(0, VEoverall, len=20)
M <- 5
alpha <- 0.05
sigma2obs <- 1
rho <- 0.9
biomType <- "continuous"
cohort <- TRUE
p <- 0.01
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             VEoverall=VEoverall, risk0=risk0, PlatVElowest=PlatVElowest, VElowest=VElowest,
             M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType, cohort=cohort, p=p)
p <- 0.02
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             VEoverall=VEoverall, risk0=risk0, PlatVElowest=PlatVElowest, VElowest=VElowest,
             M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType, cohort=cohort, p=p)
p <- 0.03
computePower(nCases=nCases, nControls=nControls, nCasesWithS=nCasesWithS,
             VEoverall=VEoverall, risk0=risk0, PlatVElowest=PlatVElowest, VElowest=VElowest,
             M=M, alpha=alpha, sigma2obs=sigma2obs, rho=rho, biomType=biomType, cohort=cohort, p=p)


## Continuous biomarker, saving output, varying sample sizes ##

nCases <- 32
nControls <- 1000
nCasesWithS <- 32
controlCaseRatio <- 5
VEoverall <- 0.75
risk0 <- 0.034
PlatVElowest <- 0.2
VElowest <- seq(0, VEoverall, len=20)
M <- 5
alpha <- 0.05
sigma2obs <- 1
rho <- c(1, 0.9, 0.7, 0.5)
biomType <- "continuous"
saveDir <- "~/myDir"
saveFile <- "MyFile"
computePower(nCases=nCases, nCasesWithS=nCasesWithS, nControls=nControls,
             controlCaseRatio=controlCaseRatio, VEoverall=VEoverall, risk0=risk0,
             PlatVElowest=PlatVElowest, VElowest=VElowest, M=M, alpha=alpha,
             sigma2obs=sigma2obs, rho=rho, biomType=biomType, saveDir=saveDir, saveFile=saveFile)
}

}
\seealso{
\code{\link{computeN}}
}
